requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
asgiref==3.7.2
    # via
    #   -r requirements/base.in
    #   django
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
django==4.2.4
    # via
    #   -r requirements/base.in
    #   djangorestframework
    #   drf-spectacular
django-environ==0.10.0
    # via -r requirements/base.in
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
drf-spectacular==0.26.4
    # via -r requirements/base.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.19.0
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
psycopg2-binary==2.9.7
    # via -r requirements/base.in
pytz==2023.3
    # via
    #   -r requirements/base.in
    #   djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.10.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via
    #   -r requirements/base.in
    #   django
typing-extensions==4.7.1
    # via
    #   -r requirements/base.in
    #   asgiref
uritemplate==4.1.1
    # via drf-spectacular
